feat:userEndpoint

+++ src/users/users.service.ts
@@ -2,7 +2,8 @@
 import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { User } from './entities/user.entity';
+import { User, UserRole } from './entities/user.entity';
+import { UpdateDto } from './dto/update-user.dto';
 
 @Injectable()
 export class UsersService {
@@ -12,7 +13,7 @@ export class UsersService {
   ) {}
 
   // Méthode create manquante
-  async create(createUserDto: User): Promise<User> {
+  async create(createUserDto: User): Promise<Omit<User, 'password'>> {
     // Vérifier si l'utilisateur existe déjà
     const existingUser = await this.findByEmail(createUserDto.email);
     if (existingUser) {
@@ -20,34 +21,91 @@ export class UsersService {
     }
 
     const user = this.usersRepository.create(createUserDto);
-    return await this.usersRepository.save(user);
+    const savedUser = await this.usersRepository.save(user);
+    const { password, ...userWithoutPassword } = savedUser;
+    return userWithoutPassword;
   }
 
-  async findAll(): Promise<User[]> {
-    return await this.usersRepository.find();
+  async findAll(): Promise<Omit<User, 'password'>[]> {
+    const users = await this.usersRepository.find();
+    return users.map(({ password, ...user }) => user);
   }
 
-  async findOne(id: string): Promise<User> {
+  async findOne(id: string): Promise<Omit<User, 'password'>> {
     const user = await this.usersRepository.findOne({ where: { id } });
-    
     if (!user) {
       throw new NotFoundException(`User with ID ${id} not found`);
     }
-    
-    return user;
+    const { password, ...userWithoutPassword } = user;
+    return userWithoutPassword;
   }
 
-  async findByEmail(email: string): Promise<User | null> {
-    return await this.usersRepository.findOne({ where: { email } });
+  async findByEmail(email: string): Promise<Omit<User, 'password'> | null> {
+    const user = await this.usersRepository.findOne({ where: { email } });
+    if (!user) return null;
+    const { password, ...userWithoutPassword } = user;
+    return userWithoutPassword;
   }
 
-  async update(id: string, updateData: Partial<User>): Promise<User> {
-    const user = await this.findOne(id);
+  async update(id: string, userDemnde: any, updateData: UpdateDto): Promise<Omit<User, 'password'>> {
+    const user = await this.usersRepository.findOne({ where: { id } });
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+    if (user.id !== userDemnde.id && !['admin', 'super_admin'].includes(userDemnde.role)) {
+      throw new NotFoundException(`You can only update your own information`);
+    }
+
+    if (!['admin', 'super_admin'].includes(user.role) && updateData.role) {
+      updateData.role = user.role; // Ignore any role change attempt by non-admins
+    }
+
+    if(user.role === 'admin' && updateData.role && ['admin', 'super_admin'].includes(updateData.role)) {
+      throw new NotFoundException(`Admins can't give themselves admin role`); 
+    }
     Object.assign(user, updateData);
-    return await this.usersRepository.save(user);
+    const savedUser = await this.usersRepository.save(user);
+    const { password, ...userWithoutPassword } = savedUser;
+    return userWithoutPassword;
   }
 
-  async remove(id: string): Promise<void> {
+  async changeStatus(id: string, userDemnde: any, isActive: boolean): Promise<Omit<User, 'password'>> {
+    const user = await this.usersRepository.findOne({ where: { id } });
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+    if (user.role === UserRole.SUPER_ADMIN) {
+      throw new NotFoundException(`You cannot change the status of a super_admin user`);
+    }
+    if(user.role === UserRole.ADMIN && userDemnde.role === UserRole.ADMIN) {
+      throw new NotFoundException(`You cannot deactivate an admin user`);
+    }
+    user.isActive = isActive;
+    const savedUser = await this.usersRepository.save(user);
+    const { password, ...userWithoutPassword } = savedUser;
+    return userWithoutPassword;
+  }
+
+  async remove(id: string, userDemnde: any): Promise<void> {
+    const user = await this.findOne(id);
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+    if (user.role === UserRole.SUPER_ADMIN) {
+      throw new NotFoundException(`You cannot delete a super_admin user`);
+    }
+    if (user.id === userDemnde.id) {
+      throw new NotFoundException(`You cannot delete your own account`);
+    }
+
+    if (!['admin', 'super_admin'].includes(userDemnde.role)) {
+      throw new NotFoundException(`You do not have permission to delete this user`);
+    }
+
+    if (userDemnde.role === UserRole.ADMIN && user.role === UserRole.ADMIN) {
+      throw new NotFoundException(`Admins cannot delete other admin users`);
+    }
+    
     const result = await this.usersRepository.delete(id);
     if (result.affected === 0) {
       throw new NotFoundException(`User with ID ${id} not found`);
